Cover separate management interface deployed by openshift-extension
Follow-up to 2b483fa2165907e5d0f18687896915ccab05e07c
Apps with separate management interface can now be deployed
by quarkus openshift-extension

Test plans for the change: https://github.com/quarkus-qe/quarkus-test-plans/blob/main/QUARKUS-2738.md
Related links:
    https://github.com/quarkusio/quarkus/pull/30506
    https://issues.redhat.com/browse/QUARKUS-2738
