+++ http/management/src/test/java/io/quarkus/qe/extension/OpenShiftExtensionIT.java
@@ -1,0 +1,8 @@
+package io.quarkus.qe.extension;
+
+import io.quarkus.test.scenarios.OpenShiftDeploymentStrategy;
+import io.quarkus.test.scenarios.OpenShiftScenario;
+
+@OpenShiftScenario(deployment = OpenShiftDeploymentStrategy.UsingOpenShiftExtension)
+public class OpenShiftExtensionIT extends OpenShiftBaseDeploymentIT {
+}
