+++ monitoring/micrometer-prometheus/src/test/resources/service-monitor.yaml
@@ -7,7 +7,7 @@ metadata:
 spec:
   endpoints:
   - interval: 30s
-    targetPort: 8080
+    targetPort: 9000
     path: /q/metrics
     scheme: http
   selector:
