+++ pom.xml
@@ -19,8 +19,7 @@
         <quarkus.platform.artifact-id>quarkus-bom</quarkus.platform.artifact-id>
         <quarkus.platform.group-id>io.quarkus</quarkus.platform.group-id>
         <quarkus.platform.version>999-SNAPSHOT</quarkus.platform.version>
-        <!-- please keep Quarkus version used in lifecycle-application module same as is used in framework -->
-        <quarkus.qe.framework.version>1.3.0.Beta13</quarkus.qe.framework.version>
+        <quarkus.qe.framework.version>1.3.0.Beta14</quarkus.qe.framework.version>
         <quarkus-qpid-jms.version>0.42.0</quarkus-qpid-jms.version>
         <quarkus-ide-config.version>2.12.1.Final</quarkus-ide-config.version>
         <apache-httpclient-fluent.version>4.5.14</apache-httpclient-fluent.version>
